Refactor Redis service to factory pattern and update imports accordingly
